% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{create_route}
\alias{create_route}
\title{Create a Route for the Given Model}
\usage{
create_route(model)
}
\arguments{
\item{model}{The model instance to be used in route handlers.
Must implement the subclass of
\href{https://github.com/eWaterCycle/bmi-r/blob/master/R/abstract-bmi.R}{AbstractBmi} # nolint: line_length_linter.}
}
\value{
A route object that can be used to interact with the model.
}
\description{
This function generates a route for the specified model. The route is used to
facilitate communication and interaction with the model.
}
\examples{
\dontrun{
model <- SomeModel$new()
route <- create_route(model)
}

}
