% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serve}
\alias{serve}
\title{Serve the BMI model}
\usage{
serve(model, port = 50051, host = "127.0.0.1", ignite = TRUE)
}
\arguments{
\item{model}{The model instance to be served. Must implement the subclass of \href{https://github.com/eWaterCycle/bmi-r/blob/master/R/abstract-bmi.R}{AbstractBmi}}

\item{port}{The port to serve the model on. Default is 50051 or if BMI_PORT environment variable is set, it will be used.}

\item{host}{The host to serve the model on. Default is "127.0.0.1".}

\item{ignite}{Whether to ignite the server immediately and block. Default is TRUE.}
}
\value{
The server application.
}
\description{
This function serves the model on a specified port and host.
}
